+++ src/br/com/projetoceetsystemsix/model/Cliente.java
@@ -1,0 +1,84 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package br.com.projetoceetsystemsix.model;
+
+import java.util.Date;
+
+/**
+ *
+ * @author Aluno
+ */
+public class Cliente {
+    private int id;
+    private String nome;
+    private String endereco;
+    private String bairro;
+    private Date clienteDesde;
+    private int telefoneCelular;
+    private String email;
+
+    
+    
+    public int getId() {
+        return id;
+    }
+
+    public void setId(int id) {
+        this.id = id;
+    }
+
+    public String getNome() {
+        return nome;
+    }
+
+    public void setNome(String nome) {
+        this.nome = nome;
+    }
+
+    public String getEndereco() {
+        return endereco;
+    }
+
+    public void setEndereco(String endereco) {
+        this.endereco = endereco;
+    }
+
+    public String getBairro() {
+        return bairro;
+    }
+
+    public void setBairro(String bairro) {
+        this.bairro = bairro;
+    }
+
+    public Date getClienteDesde() {
+        return clienteDesde;
+    }
+
+    public void setClienteDesde(Date clienteDesde) {
+        this.clienteDesde = clienteDesde;
+    }
+
+    public int getTelefoneCelular() {
+        return telefoneCelular;
+    }
+
+    public void setTelefoneCelular(int telefoneCelular) {
+        this.telefoneCelular = telefoneCelular;
+    }
+
+    public String getEmail() {
+        return email;
+    }
+
+    public void setEmail(String email) {
+        this.email = email;
+    }
+    
+    
+    
+    
+}
