+++ src/br/com/projetoceetsystemsix/model/ClienteFisico.java
@@ -1,0 +1,34 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package br.com.projetoceetsystemsix.model;
+
+/**
+ *
+ * @author Aluno
+ */
+public class ClienteFisico extends Cliente{
+   private int CPF;
+   private int RG; 
+
+    public int getCPF() {
+        return CPF;
+    }
+
+    public void setCPF(int CPF) {
+        this.CPF = CPF;
+    }
+
+    public int getRG() {
+        return RG;
+    }
+
+    public void setRG(int RG) {
+        this.RG = RG;
+    }
+    
+   
+ 
+}
