+++ src/br/com/projetoceetsystemsix/model/Compra.java
@@ -1,0 +1,42 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package br.com.projetoceetsystemsix.model;
+
+/**
+ *
+ * @author Aluno
+ */
+public class Compra {
+    private int id;
+    private int quantidade;
+    private double valorCompra; 
+    private Fornecedor fornecedorCompra; // ultimo feito
+    public int getId() {
+        return id;
+    }
+
+    public void setId(int id) {
+        this.id = id;
+    }
+
+    public int getQuantidade() {
+        return quantidade;
+    }
+
+    public void setQuantidade(int quantidade) {
+        this.quantidade = quantidade;
+    }
+
+    public double getValorCompra() {
+        return valorCompra;
+    }
+
+    public void setValorCompra(double valorCompra) {
+        this.valorCompra = valorCompra;
+    }
+     
+    
+}
