+++ src/br/com/projetoceetsystemsix/model/Fornecedor.java
@@ -1,0 +1,123 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package br.com.projetoceetsystemsix.model;
+
+/**
+ *
+ * @author Aluno
+ */
+public class Fornecedor {
+    private int id;
+    private String razaoSocial;
+    private int CNPJ;
+    private String logradouro;
+    private String complemento;
+    private String bairro;
+    private String cidade;
+    private String UF;
+    private int Cep;
+    private int telefoneComercial;
+    private int telefoneCelular;
+    private String email;
+
+    public int getId() {
+        return id;
+    }
+
+    public void setId(int id) {
+        this.id = id;
+    }
+
+    public String getRazaoSocial() {
+        return razaoSocial;
+    }
+
+    public void setRazaoSocial(String razaoSocial) {
+        this.razaoSocial = razaoSocial;
+    }
+
+    public int getCNPJ() {
+        return CNPJ;
+    }
+
+    public void setCNPJ(int CNPJ) {
+        this.CNPJ = CNPJ;
+    }
+
+    public String getLogradouro() {
+        return logradouro;
+    }
+
+    public void setLogradouro(String logradouro) {
+        this.logradouro = logradouro;
+    }
+
+    public String getComplemento() {
+        return complemento;
+    }
+
+    public void setComlemento(String complemento) {
+        this.complemento = complemento;
+    }
+
+    public String getBairro() {
+        return bairro;
+    }
+
+    public void setBairro(String bairro) {
+        this.bairro = bairro;
+    }
+
+    public String getCidade() {
+        return cidade;
+    }
+
+    public void setCidade(String cidade) {
+        this.cidade = cidade;
+    }
+
+    public String getUF() {
+        return UF;
+    }
+
+    public void setUF(String UF) {
+        this.UF = UF;
+    }
+
+    public int getCep() {
+        return Cep;
+    }
+
+    public void setCep(int Cep) {
+        this.Cep = Cep;
+    }
+
+    public int getTelefoneComercial() {
+        return telefoneComercial;
+    }
+
+    public void setTelefoneComercial(int telefoneComercial) {
+        this.telefoneComercial = telefoneComercial;
+    }
+
+    public int getTelefoneCelular() {
+        return telefoneCelular;
+    }
+
+    public void setTelefoneCelular(int telefoneCelular) {
+        this.telefoneCelular = telefoneCelular;
+    }
+
+    public String getEmail() {
+        return email;
+    }
+
+    public void setEmail(String email) {
+        this.email = email;
+    }
+ 
+    
+}
