+++ src/br/com/projetoceetsystemsix/view/Cadastro.form
@@ -96,11 +96,6 @@
                   <Color blue="0" green="0" red="a1" type="rgb"/>
                 </Property>
               </Properties>
-              <AccessibilityProperties>
-                <Property name="AccessibleContext.accessibleParent" type="javax.accessibility.Accessible" editor="org.netbeans.modules.form.ComponentChooserEditor">
-                  <ComponentRef name="CadastroFornecedor"/>
-                </Property>
-              </AccessibilityProperties>
 
               <Layout class="org.netbeans.modules.form.compat2.layouts.support.JTabbedPaneSupportLayout"/>
               <SubComponents>
