+++ src/br/com/projetoceetsystemsix/view/PesqProduto.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package projetoceetsystemsix;
+package br.com.projetoceetsystemsix.view;
 
 /**
  *
